Added documentation for some of the internal routines; Reverted to using uint32_t instead of uintptr_t; Added checks for error when retrieving the mode, pull and speed values;

+++ cc_hal/pin_defs.h
@@ -270,7 +270,7 @@ typedef struct {
 	pin_name_t pin_name;	/**< Associated pin name from the above list */
 	gpio_config_t settings;	/**< GPIO pin configuration to go with peripheral */
 	uint8_t alt_func;	/**< Alternate function ID for the pin */
-	uintptr_t peripheral;	/**< Associated peripheral, if any */
+	uint32_t peripheral;	/**< Associated peripheral, if any */
 } pin_t;
 
 /**
